refactoring to separate hetero layer
and minor improvements
